+++ example/lib/demos/app_bar.dart
@@ -1,4 +1,4 @@
-import 'package:cupertino_native_plus/cupertino_native.dart';
+import 'package:cupertino_native_better/cupertino_native_better.dart';
 import 'package:flutter/cupertino.dart';
 
 class AppBarDemoPage extends StatelessWidget {