+++ example/lib/demos/icon.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import 'package:cupertino_native_plus/cupertino_native.dart';
+import 'package:cupertino_native_better/cupertino_native_better.dart';
 
 class IconDemoPage extends StatefulWidget {
   const IconDemoPage({super.key});