+++ example/lib/demos/image_asset_demo.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
-import 'package:cupertino_native_plus/cupertino_native.dart';
+import 'package:cupertino_native_better/cupertino_native_better.dart';
 
 class ImageAssetDemo extends StatefulWidget {
   const ImageAssetDemo({super.key});