+++ test/cupertino_native_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:cupertino_native_plus/cupertino_native.dart';
+import 'package:cupertino_native_better/cupertino_native_better.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 class MockCupertinoNativePlatform